+++ Raytracer/Vector2F.cpp
@@ -4,7 +4,7 @@ Vector2F::Vector2F() : X(0), Y(0) { }
 Vector2F::Vector2F(const float& x, const float& y) : X(x), Y(y) { }
 Vector2F::Vector2F(const float vec[]) { X = vec[0]; Y = vec[1]; }
 
-float Vector2F::Length() { return sqrt(LengthSquared()); }
+float Vector2F::Length() { return sqrtf(LengthSquared()); }
 float Vector2F::LengthSquared() { return Dot(*this, *this); }
 
 float Vector2F::Dot(const Vector2F& other) const { return Dot(*this, other); }
